Make sure public functions parameters are unaffected by argument promotion
In order for a prototype declaration and non-prototype function
definition to be compatible, the promoted type of the non-prototype
function parameters must be compatible with the type of the corresponding
prototype parameter.

Since funcs.h is autogenerated by mkfuncs.pl which copies the parameter
types as-is, use the promoted type in the definitions so that the types
are unaffected by default argument promotion.

Fixes #11