Remove redundant preprocessing of the less.hlp file
There are no placeholders in less.hlp.VER, so it's pointless to preprocess
it. Removing this step simplifies the build process a little.

This also fixes an issue with Makefile.wng: if you were building from
a release tarball and ran "make clean", you wouldn't be able to rebuild,
since less.hlp.VER isn't included in release tarballs.