Avoid spurious empty lines before the beginning of the file (#93)
When going to the special line 0 (default for 'g' / '<'), ignore
jump_sline in order to avoid empty lines before the beginning of
the file. Note that jump_sline is still honored when going to a
positive line number, so that empty lines may appear in this case
(like in searches).