Fix "Tag not found" error while looking for a tag's location
In findctag() restore code to actually remove '\' from the pattern
buffer. The comment in the function says it is removing backslashes,
but that code was removed during edits around 12/11/2001.  This fix
restores the original code except that it avoids copying characters
on top of themselves.  Instead, it only begins to copy characters
after encountering a backslash.