Fix crash due to memcpy with overlapping objects
According to the POSIX specification of memcpy(3). If a copy takes place
between objects that overlap, the behaviour is undefined. Unfortunately,
horizontal scrolling contained two such memcpy invocations which lead to
crash on Alpine Linux. Since The _FORTIFY_SOURCE implementation used on
Alpine Linux, checks for memcpy invocations with overlapping objects and
crashes the program when such an invocation is found. On Alpine this
crash can be reproduced by opening a file which exceeds the terminal
width in the horizontal dimension and then attempting to scroll
horizontally.

This commit fixes this issue by using memmove(3) instead of memcpy(3),
thereby fixing a regression introduced in 143ba345 and d7f6d4ec.