Fix memory leaks
If LESS or MORE environment variable is set, then a copy of the value
is created for scan_option call. Parse value directly without creating
a copy since the allocated string is not modified in scan_option.

Also handle rare cases in which a history file has duplicated mark
entries.