windows: mingw: Makefile.wng: support REGEX_PACKAGE=none
The C source code supported NO_REGEX since configurable regex was
introduced at 55b5b97, but it was not possible to use it because
in such case Makefile.wng still tried to enable some lib/objects
which might not be available.

Now REFEX_PACKAGE=none works and lets defines.wn set NO_REGEX=1.

This is likely useful only for testing, or when size really matters.