Port to Autoconf 2.71 (#320)
This consisted of running 'autoupdate' and then minimally
fixing up to keep the patch small and the build working.
The manual fixes were mostly just to fix indenting.
However, one nontrivial fix was to stop using RETSIGTYPE
which Autoconf 2.63 (2008) declared permanently obsolete.
The only platform that cares about this is the OS-9 version for
Microware C 3.2, and I expect the code will still compile and run
anyway (perhaps with a warning), in the unlikely case that an OS-9
user builds with Makefile.o9c instead of Makefile.o9g or Makefile.o9u.