Windows/MS-DOS: decouple SGR state from screen colors
The trigger for this commit was to make the -D<x> mapping more
predictable. Instead of modifying win_flush (~250 LOC), the SGR
and colors logic was refactored out of win_flush, which was an
opportunity to make the logic simpler and more rebust, resulting in:

1. Apply the -D{d,s,k,u} mapping more consistently and predictably.

Previously, the -D<x> values where applied as follows:

By default (i.e. without -Da), if an SGR sequence was a single
attribute-related number, then the color was set as follows,
regardless of the attribute or the operation (set/unset):

  If BOLD is set, use the -Dd value,
  Else if INVERSE and/or ITALIC were set, use -Ds,
  Else... <same for UNDERLINE and BLINK with -Du, -Dk>
  Else, use default colors

This resulted in stateful and unpredictable application of the -D<x>
values depending on order of values and/or SGR sequence length, e.g:

  \e[0;1m            did not apply any -D value
  \e[0;33;1m         no -D value
  \e[0m\e[33m\e[1m   used -Dd (bold) value
  \e[0m\e[1m\e[33m   no -D value
  \e[0m\e[1m         used -Dd (bold) value
  \e[0m\e[1m\e[7m    also used the -Dd value... (should be -Ds ?)
  \e[0m\e[41m\e[22m  default colors instead of red bg (no attributes)

The -Da behavior ("SGR mode") had neither the -D<x> mapping nor the
stateful behavior, which resulted in much more standard and consistent
application of SGR sequences. This is still the case now.

But without -Da, the -D<x> values are now applied whenever the fg+bg
are default and exactly the relevant -D<x> attribute is set, regardless
of how it got to this state. Except for this reasonably predictable
state mapping, it's now identical to, and as consistent as "SGR mode".

This should have little negative impact in practice, if at all, because
the current behavior was likely the main use case of -D<x> in the past
too (e.g. manpage), but now it's consistent and predictable by default.

However, surely there are users which will consider this a regresion...

Note that because italic is applied as inverse (standout), previously
-Ds (standout) affected italic too. This behavior is explicitly
preserved for backward compatibility.

2. Previously, italic/inverse clashed - they shared the same bit,
   and blink/underline clashed both on their own and with the
   fg/bg/intensity - because they were not maintained independently.

E.g. \e[3;27m (set italic, unset inverse): 27 wrongly canceled 3.

Now all the attributes are maintained independently, and are applied
to Windows colors in one stateless and consistent step.

3. Additional small changes:
- Handle SGR 25 too (blink off).
- Don't treat unknown SGR values as if they were 0, e.g. \e[10m .
- Remove the difference between \e[0m and \e[0;...m (non standard).
- COMMON_LVB_UNDERSCORE no longer needs to be considered explicitly.

Potential further improvements:
- Add -Di to affect italic independently of inverse.
- Ignore or apply SGR 38/48 (256, true colors) which have extra params
  (currently 38/48 are ignored, but their additional params are not).