mouse: bind right-click as jump to the "#" mark
Previously any mouse-button click marked the line with '#'.

Now left-click marks with '#', and right-click jumps to this mark.

In practice, and to support 2/3 buttons mice, we apply the jump
with any mouse button which is not X11MOUSE_BUTTON1 (left).

With the legacy x11 mouse report handler, a button-release event
doesn't have button ID, so to trigger on a specific button release,
a state was introduced which holds the last button-press.