configure.ac: Check for tinfow before tinfo
We currently check for ncursesw > ncurses and then tinfo > tinfow. This means
we can get a mismatch of ncursesw + tinfo, instead of the correct ncursesw + tinfow.

Swap the order so we check for ncursesw first (before other ncurses variants)
and then tinfow first (before other tinfo variants).

This is needed anyway for correctness, but also needed for certain terminfos
to work correctly with recent ncurses.

A better fix would be to use pkg-config first which handles this correctly
and would include the appropriate -ltinfo* in the libraries list for -lncurses*,
but not doing that for now.

Bug: https://bugs.gentoo.org/910430