windows: mingw: Makefile.wng: use mkhelp.sh if perl is missing
On Windows perl can be installed, but it's a non-trivial dependency.

If perl is missing to process less.hlp, then try also the newly added
mkhelp.sh script (which uses od and awk in standard POSIX syntax).

The makefile already assumes grep and sed (for funcs.h), and it's
highly likely that sh, od, and awk are available as well.

The minimum requirements are now a mingw compiler, gnu make, sh or
cmd.exe, and gnu coreutils or busybox[-w32] - on *nix or Windows.