windows: mingw and old msvc: don't use _snprintf
Previously, msvc before 2015 and mingw defined snprintf as _snprintf.

This can be a problem because _snprintf is non-compliant and does not
null-terminate if the length exceeds the buffer.

Also, some non-latest mingw versions don't have _snprintf (gcc 10).

With mingw this was not needed at all, because mingw ensures there's
a compliant snprintf (both with msvcrt and with ucrt), so now mingw
uses the standard snprintf.

With msvc before 2015 there's indeed an issue, but now instead of
using the broken _snprintf we specify that we don't have snprintf,
so that the appropriate code paths are chosen (currently that uses
sprintf, but in the future there could be a better shim).