windows: mingw: Makefile.wng: default to regcomp-local
Previously, the default REGEX_PACKAGE was posix, but this is not
available by default in most/all mingw-w64 setups (debian, fedora,
w64devkit, and more).

Additionally, the regex default prefix is /mingw-w64/mingw64/opt
(when using posix or gnu regex), which is not available in most
mingw setups (not MSYS2 mingw 64 bits, and not even cygwin), so
this mostly didn't work.

In contrast, regcomp-local works in any mingw setup and doesn't
require non-standard libs, so that's a safer default.

Also, the non-mingw default is regcomp-local, and that's also used
at the semi-official windows builds of less, so it's better to have
the same default on all windows builds.