windows: mingw: support files bigger than 2G
The changes of commit 44d5c6bb can also work for mingw, so enable it.

The type _int64 is not identified in mingw without further includes,
so use long long instead (which is synonymous to __int64 according
to MS docs)

Tested 32 and 64 bit mingw builds, with the 32 bits confirmed to open
and seek to the end of a 7G file (where previously it didn't), and
also tested and works on XP (tested only small files on XP).