windows: msvc: Makefile.wnm: allow generating funcs.h, help.c
Previously the mscv Makefile.wnm didn't have rules to generate funcs.h
and help.c, meaning that it could only be used to build release
tarballs (which already have these files), but not git snapshots.

Now we have a new independent rule "generated" (not part of "all"),
which generates them. Because typical msvc setups do't have the tools
which are used elsewhere to generate these files, this commit adds
buildgen.c which, when compiled, can be used instead.

It's not part of "all" because when cross compiling, e.g. an arm
binary on Windows 64, buildgen.exe should be native and not arm,
so this allows making "generated" with a native toolchain, and then
"all" using the arm toolchain.