Windows: mingw: Makefile.wng: allow native funcs.h, help.c
We already have buildgen.c which can be used to generate funcs.h and
help.c without grep/sed/perl/sh/etc, which can be useful in mingw
setups on Windows which don't bring the whole posix arsenal with them.

The default is still the posix tools, but now adding WINGEN=1 will
use buildgen.c instead of the posix tools.

The compiler which is used with buildgen.c is the same compiler used
at the rest of Makefile.wng, so it should be native (not cross build).