windows: fix console title with unicode file name
Since commit ad440b6, we can open unicode file names as UTF-8 if the
binary was built with UCRT, however, because this is a runtime
attempt, the build-time CP_ACP might be incorrect, and instead,
less_acp should be used to take into account successful UTF-8 locale,
as noted in a comment in that commit.

Do that for the console (window) title, which un-gibberish the text
with unicode file names.