windows: improve UTF-8 locale comment (no-op)
Commit ad440b6 added support for runtime UTF-8 locale, but incorrectly
documented that it's quite the same as a UTF-8 manifest, except
startup values which don't automatically become UTF-8 (like argv).

However, that's not quite true.

While the UTF-8 manifest affects all the A(NSI) Windows APIs, like
CreateFileA and fopen, the runtime setlocale method affects only
the libc/crt API, and so fopen is covered, but CreateFileA is not.

Fix the comment to mention that.

Currently "less" uses these non-libc ANSI Windows API:
- CreateFileA: only used with the ASCII name "CONIN$", so OK..
- FillConsoleOutputCharacterA: only used with spaces and other ASCII.
- ScrollConsoleScreenBufferA: same as above.
- WriteConsoleA: only used when the console CP is UTF8, so OK.
- {Get,Set}ConsoleTitleA: saves/restores the original title, and uses
  console CP regardless of manifest/locale. Will be fixed later.