windows: restore unicode console title correctly
"less" saves the console title on startup, and restores it on exit,
however, it does so using the A(NSI) win API, and so if the title
was originally unicode, then restoration may fail, depending on the
console and ANSI codepages and other factors.

Now the save and restore are with the W(ide) win API, so it doesn't
depend on any codepages.

This is unrelated to unicode file names, and works even when "less"
doesn't support unicode file names.