When not using -Da (and always on Windows before 10), we process SGR sequences internally and apply them as Windows/DOS colors.
Previously, for SGR codes of 50 or more (e.g. ^[[38;2;150;200;250m)
the SGR processor considered it invalid sequence, and passed-through
this sequence and the rest of the current buffer (typically line).

This resulted in 256 or true color sequences passed to the terminal,
and on terminals which don't support sequences natively (i.e. before
Windows 10), this resulted in incorrect output - the sequence prints
literally.

This commit does two things:
1. Identify unknown codes exactly, rather than "code >= 50"
   (e.g. 38 is unknown, while 39 or 40 are known and supported).
2. On unknown code, don't apply more codes at this sequence.

This fixes incorrect junk output for 256/true colors on Windows
earlier than 10, e.g. as part of the content (with the -R option),
or as part of 8-bit color transformation with the -Dx options.

256/true colors still don't apply (i.e. they're ignored), but now
such sequences don't print junk.

Note that there are still sequences which the SGR processor considers
invalid, such as various OSC sequences, and more, and on these cases
the old behavior of "passthrough till the end of buffer" still apply.