windows 10+: make 256/true colors "work" without -Da
"sgr mode" (-Da) bypass the internal SGR processing and passthrough
everything to the terminal directly, which works fine on win10+
where the terminal supports more sequences than us - like 256 colors.

But by default (without -Da), we do internal SGR processing and apply
the values to the console outselves, including some transformation of
-D<x> values unique to windows (like underline to cyan).

Before commit f9f2498f (which is just before v651), unknown SGR
values resulted in aborting the SGR parsing and instead passthrough
till the rest of the buffer to the terminal directly.

This had broken output if the terminal doesn't support VT (e.g. win7),
but seemingly correct output with VT, e.g. 256 colors on win10.

It did have some issues still even with VT, like losing track of the
terminal state, so if the next buffer has e.g. SGR 1 (bold) then we'd
output a wrong color (the last state we parsed, plus bold).

It also stopped applying the various -D<x> config values till the end
of this buffer, because this buffer was switched to passthrough.

Commit f9f2498f changed this process, by ignoring only the rest of the
current sequence if an unknown value is encountered, so that the next
sequences in the buffer are still processed normally.

Additionally, we then never lose track of the terminal state, because
either we recognize a value and apply it, or we do neither of those.
Specifically, we never apply a value which we can't recognize/track.

While it fixed the junk output when VT is disabled, it also had an
effect which made 256 and true colors stop "working" on win10+.

So this commit brings back this behavior of "working" 256/true colors
on win10+ by default, but a bit more refined than before f9f2498f .
I.e. in v643 it appeared that 256 colors work by default, but in v661
it appeared to have been broken, and now it's fixed.

When VT is disabled it's now the same as v661, and like in v661,
it also doesn't have junk output from unknown sgr values.

When VT is enabled and an unsupported SGR value is encountered, we
stop applying the SGR values just like in v661, and also keep parsing
just like in v661.

However, instead of applying the values which we did recognize like
in v661, we now switch to a mode which passthrough the whole sequence,
because the terminal can handle it, but we know we've lost sync.

Then, we keep parsing sequences for hints but also keep passthrough
the sequences until we know what the terminal state is - after reset.

That is, the scope of "unknown terminal state and no -D<x>" is now
minimal and exact (till first reset), instead of till the first
reset at the next buffer or later, and also we don't get hit by wrong
color when we've lost track with the terminal state.

This difference it non negligible, because many times the buffer is
a full line - which ends in a reset. So now we identify this reset
so that the next line is in sync, but in v643 the sgr state at the
next line was the one we had before we passthrough the rest of the
buffer, so the whole next line could be wrong.

Fixes #538