windows: SGR processor: improve end-of-buffer detection
The code at win_flush expects is_ansi_end(c) to return 0 if c == 0,
but it actually returns 1.

Due to that, it could sometimes fail to detect the end of the input
buffer (which ends in \0), read past it, and sometimes crashed.

Reaching the end of the buffer (1K) was rare - only when the visible
part of an output line is generated from more than 1K bytes.

This doesn't necessarily mean that the visible part of the line spans
more than 1024 columns (although this would also be such a case), but
rather that there are many invisible bytes at the line, like escape
sequences, so that e.g. 80 or 100 columns output can be generated
from 1K string which includes many SGR escape sequences.

This commit adds a new function is_ansi_end_0 which behaves like
win_flush expects is_ansi_end to behave, and now uses it instead.

During testing with a tiny buffer (8 bytes) and many SGR escape
sequences (not bigger than 8 bytes) which trigger end-of-buffer
behavior many times for each line, no bad behavior was observed.

There remains at least one non-critical issue where if a sequence
begins at the last byte of the buffer, then the ESC is passthrough
instead of being processed as SGR, but it's un-noticeable on win10+
because the terminal handles the ESC sequence in such case.