Config files, like _lesskey, are determined at homefile(...), and on windows, after trying $HOME and $HOMEDRIVE$HOMEPATH, it could fallback to using _searchenv, which first searches CWD and then in $PATH.
Disable this fallback on Windows.

Note that OS2 and DOS variants (DJGPP, Borland, MS-DOS-compiler) still behave like before, and search PWD and $PATH (assuming they still compile - all the non-DJGPP targets use and used _searchenv, which probably only exists on WIndows).

At the very least the DOS variants don't have a notion of $HOME, so it's not obvious where the config files should be searched - if we wanted to address these build targets as well, but for now we don't.

Manual merge of #527.