Fix --shift and --match-shift on command line with fractional parameter.
Setting --shift calls calc_shift_count() and setting --match-shift calls
calc_match_shift(). When the parameter begins with a '.', these calculate
the value as a fraction of the screen width. But at startup when these
functions are called from scan_option(), the screen width is not yet known.
So call them in get_term() after the screen width is first determined.
Previously this was done for calc_jump_sline() but it also needs to be
done for the other two.