Fix bug related to ctrl-X when output is not a terminal.
We should not check for ctrl-X input when the output is not
a terminal. This results in trying to poll and read input
characters from stdin instead of from the terminal.

This problem existed in v661 so is not strictly a regression
that needs to be addressed in the upcoming bug fix release,
but because it's a low-risk fix and has potentially serious
consequences, I'm including it.

Related to #558.