Fix Windows build.
There was a conflict between the definition of BOOL in option.h
and a BOOL typedef in a Windows header file. Move the #include of
option.h after the #include of the Windows headers.
This is rather ugly. We should probably rename the #defines in
option.h to avoid such conflicts.
Related to #561.