Fix problem in getccu when ctrl-Z is entered.
getccu calls getchr which calls iread.
When ctrl-Z (SIGTSTP) is entered, iread sets sigs |= S_STOP
and returns READ_INTR. getchr then also returns READ_INTR.
As of 2d0abe8a, getccu converts any negative value like READ_INTR
to a null char, which causes getccu to ignore the char and call
getchr again. Before 2d0abe8a we would retain the negative char
and exit the loop, but the char value was not used because the
fact that sigs is set would be checked first. It seems more
straightforward to just exit the getccu loop when sigs is set.

Related to #563.