Add warning when F command is used with LESSOPEN.
Normally when a LESSOPEN script is used, the script reads the whole
input file, transforms it, writes the result to the replacement
file or a pipe into less, and then exits. Any changes to the input
file after that point will not be seen because the script has
terminated. This is a warning rather than an error because it is
conceivable that a LESSOPEN script could be written to remain
active after reading EOF, watching for further changes to the file,
although this is not the way the vast majority of scripts are written.

Related to #291.