Allow reading from a fifo that is closed and reopened by the writer.
9933cf193898275b3a6209029d99e1ef5482fa6c made changes so that after
reading to EOF, if more data is appened to the file, scrolling forward
at the old EOF would pick up the new data. However this did not work
for pipes/fifos, since the change uses filesize() to detect that new
data has appeared. This change allows reading new data in a fifo as well.