Fix use of multibyte characters as commands.
When we process the first byte(s) of a multibyte command char,
we should just accept it and get the next byte, without changing
state. The previous code was calling cmd_reset(), which discarded
any previous chars in the cmd buffer.

This was probably broken in 7e6da25e345b146e2ced95a705dfe865d8c650cc.

Related to #595.