Fix background color problem.
Writing a colored char into the last position of a line on auto_wrap
terminals can cause the whole next line to get colored. Sending a
color reset sequence at that point will not change the color of that
line. Send a clear_eol sequence after resetting color, to get rid
of the erroneous line coloring.