Don't poll input file if -F will exit.
When -F is in effect and the file is short, so less will exit
after displaying the first screen, don't poll the input file.
Polling eats tty input that the user might expect to be directed
to the shell if they know that the input file is short.

Related to #599.