Make SIGTSTP work as expected if received during iopen.
In ca59eda0baea2409ebef33406269474d61efd93b iopen was created
which handles SIGINT received during a blocked open(). But as
noted in the commit message, SIGTSTP was handled the same as SIGINT.
This commit fixes SIGTSTP to stop the process rather than kill it.
This required changing psignals() to send SIGSTOP rather than SIGTSTP
to the current process to stop it. SIGTSTP doesn't stop the process
even though the handler is set to SIG_DFL. I don't know why, but
it seems harmless enough to use SIGSTOP rather than SIGTSTP.