Treat -r in LESS environment variable as if it were -R.
The -r option is very rarely useful, and can even be harmful,
in that a file containing lines longer than the screen
width will be displayed incorrectly. Also, viewing a file
containing escape sequences could put the terminal in an
unknown or unsafe condition.  The user is expected
to use -r judiciously, invoking it only when the viewed file
is known to contain "safe" content.

Therefore, it is almost always an error to put -r in the LESS
environment variable.  In all but exceedingly rare situations,
the -R option should be used instead.

This commit makes less treat any -r option in a LESS environment
variable as if it were -R. Intentional use of -r must be done
on the command line or via the - command.