New strategy for handling highlighting.
Previously, prep_hilite was called in several places and would try to
prepare a largish section of the file at once. The size of the prepped
section was more or less random, depending on the size of the line buffer
at the time of the call. This was ostensibly done for "efficiency", but
it's not really much more efficient than prepping smaller sections in
several calls to prep_hilite. It has the negative side effect that
when file input was not immediately available (eg. a slow pipe), less
would get stuck in prep_hilite waiting for data, and not display data
that was already available.

This commit changes the strategy so that we generally prep only one line at
a time. This avoids unnecessarily waiting for more data before displaying
data that is already available. There are also some changes specifically
to improve the behavior of the & command, especially when dealing with
slow input.

Related to #532 and others.