lesstest: remake seq200 test.
This test fails after change da2a9ecdf16beb642d0c030e35f0351c5f2e5a12
because we now display END in the prompt only at the actual end of
the file. If & filtering is in effect and we reach the last
unfiltered line, we don't display END unless that line is actually
the last line of the file. Retaining the old behavior would require
doing an extra unnecessary read after displaying the last unfiltered
line, which might block unnecessarily if reading a pipe. The new
behavior is also arguably more consistent; for example when filtering
we don't renumber the lines; they retain the line numbers of the
unfiltered file.