Add soft_eof to record position of forw_line returning NULL_POSITION.
This allows the prompt to display END when the user attempts to scroll
past the last line while & filtering is in effect. It almost restores
the behavior changed by da2a9ecdf16beb642d0c030e35f0351c5f2e5a12 but
not quite: the old code would display END when the last unfiltered
line was reached, but now it requires the user to attempt to move
forward from that line. I think this is an acceptabl compromise to
restore most of the old behavior without reverting to the behavior
that would sometimes hang waiting for new data without displaying
data that had already been read.